add (mflux) : configurator
Added a new configurator module to mflux that will handle parsing of json config file. A configurator object will only be made once as a read-only entity, thereby limiting the amount of times the config file is parsed to only one. It will be then shared throughout the application with other modules, components, as required.

Renamed the workspace from 'mflux' to 'mflux-app' so that it is now distinguishable between itself and the 'mflux' target that compiles as a library.

Removed the boilerplate module 'moda'