update (mqtt) : message
Implemented basic functionality of mqtt message except its serialization and deserialization into and from a string. Those are the planned features that will be later implemented 'if' necessary.

The definition is inlined into the header file as most of the methods are pure getters and setters that do not exceed more than 4-5 lines of code.