fix (mflux) : engine is running
Correct evaluation of engine's running status. Earlier, engine was said to be running only if the Core was running.
But engine comprises of two other modules, Influx and MQTT service client, as well. Now the engine is marked as running if any of these modules is found to b running.

This will ensure in proper shutting down of the system as well as we'll now know when other modules didn't stop.