update (mflux) : Send buffer reference to clients
Passed a reference of buffer to both MQTT and Influx clients so that they can push or pull messages from it as they please.

However, this has created a cyclic dependency in modules as MemDB depends on Influx and MQTT for their message types meanwhile MQTT and Influx clients depend on MemDB for buffer based operations.

With correct usage of templates, this cyclic dependency can be broken as MemDB shouldn't actually need to depend on any module for its data type. It should work without any type restrictions and a database should operate only on fundamental data types. One solution to this would be the use of templates, serialized messages or individual values.