fix (memdb) : automatic creation of new column
Whenever a push command is made, the database first checks whether it has a column present to push the data into. If it doesn't, no push is made.

This isn't ideal as during program's startup, the database is basically empty and contains no column. So whenever a push command is made, it will try find an existing column to push into. But since there isn't one. It won't push anything and the database stays empty.

So in this fix, we keep checking the existence of a column, if it's present then we push into it like normal. If a column doesn't exist, we create a new column right there and then push the newly received data into it.