fix (mflux) : incorrect translations
Earlier any received mqtt message wasn't properly getting transferred to influx, say a MQTT message arrived with value 'VAL', it would then get translated correctly to an Influx message that will contain value 'VAL'. But in a next event when, say, a new MQTT message arrived with a new value. say. 'ABC', it then doesn't get correctly translated into a Influx message. The translated Influx message would still contain the value 'VAL' that was stored in it from initial translation instead of the expected 'ABC'.

So it was evident, the translation was being done correctly, the assignment and updating of the Influx message was at fault.

In the minix module, we retrieve values from mqtt message and put it into a influx message. Note here that, we are making changes on a reference of an existing influx message, not a fresh new copy, that contains values from previous translation.

To do a successful overwrite of new values, the influx message needed to be cleared from any existing measurements in it. And then updated with new values.

Mflux, as a structure, wasn't at fault. All threads, assignments, copying, passing is being done correctly, as expected in the first attempt. Sol, the mflux is evaluating to be completely valid in its design verification.

This assignment issue was a minor inconvenience that has been fixed by adding a new method to influx::Message class as clear()