fix (influx) : message getters and docs
The commented documentation has now been appended with /// to make it compliant with doxygen toolkit that will be used to create developer guide later on.

Added const keywords to getters to ensure they do not accidentally edit the values of the object during get() operation. This also helped in catching a potential bug in Message::get_tag_value where square brackets were being used to get a value for a tag instead of the ideal ::at() function that doesn't allows bi-directional assignment.