fix (mflux) : safe shutdown
Fixed erroneous termination of Engine when it fails to start correctly. e.g; The engine currently fires up three parallel units when started. If any of the unit fails to start up due to any reason, the engine is marked as failed to start. But it still has other 2 units correctly running on their separate threads. To terminate the mflux correctly and safely, it does need to close those units as well. This condition wasn't addressed in previous builds, but now it is fixed.

Added concise logging to existing classes and modules. WIP.

Updated TODO for future refactor issues.