fix (mflux) : nullptr exception
If mflux is started without a configuration file then during cmd args parsing, Mflux doesn't creates a Configurator object and returns an error. This in turn leads to not initialization of the Engine object. This Engine object was then referenced nearing the end of Mflux::start() method. But since the Engine object wasn't initialized in the first place. It cannot be referenced to make it stop. Which leads the application to throw a nullptr exception.

This error has been handled now by checks whether the Engine object is initialized or not before requesting to stop it.