fix (mflux) : excess verbosity
The command args option -v is provided to switch the console logging level of Mflux from INFO to DEBUG. Where DEBUG being more verbose.

However, it was currently being set to be TRACE that logs out most voluminous output and is certainly not meant to be used for console logging. At max console logging is limited to just DEBUG.

Fixed the assignment of logging level for console log here.