fix (influx) : fail to find httplib.h
The influx client.h was unable to find httplib/httplib.h during linking stage of application build. Technically, it should work the way it currently is, HTTPLIB being privately linked to INFLUX module. But there is something that I do not understand in CMake, this requires me to change the directory include type for HTTPLIB from PRIVATE to PUBLIC